move options to tools
